Factor out the alter state debug component
