+++ rlbot_gui/gui/js/story-alter-save-state.js
@@ -1,0 +1,31 @@
+
+export default {
+    name: 'alter-save-state',
+    props: { 'value': Object },
+    template: `
+    <div>
+    <b-button id="debug-state-target">
+        Alter State
+    </b-button>
+    <b-popover target="debug-state-target">
+        <template v-slot:title>
+        Update UI state
+        </template>
+            <textarea cols="40" v-bind:value="JSON.stringify(value, null, 2)" v-on:input="handleInput"></textarea>
+            <b-button @click="sendJSON">Update</textarea>
+    </b-popover>
+    </div>
+    `,
+    data: {
+        "text": ''
+    },
+    methods: {
+        handleInput: function (event) {
+            this.text = event.target.value;
+        },
+        sendJSON: function () {
+            console.log(this.text)
+            this.$emit('input', JSON.parse(this.text))
+        }
+    }
+};
